Add files via upload
This program generates a random password based on varying levels of password strength as well as constraints which are decided by the user. There are 3 strengths and with each strength the password incorporates more complex characters such as both cases of letters, numbers from 0-9 as well as symbols. It also asks the user how long they want to make the password and even saves their password history in a .txt file which the user can access in the main menu of the program. 

This program uses the concepts I used in my first python project "rock, paper, scissors" but adds on to it by using list comprehensions and file managing as well.